Back to chars until the very end.
@katietz, please take this on if you can. It is very late now for me
I'll try to get back in touch as early as I can.

We need the Python code to convert from whatever string type its passing
to bytes encoded as utf8 (conda has lots of reference for such things).

Then the lowest level functions will convert them to wchar_t we pass that
to the libarchive _w API functions (2 of them).

In theory only the Python bits needs to fixed at this point, but we also
need to make sure that when libarchive is built it's built with UNICODE
compiler flag set